requirements.txt
boto3==1.37.3
colorlog==6.9.0
dvc==3.59.2
h5py==3.13.0
hydra-core==1.3.2
jinja2==3.1.6
joblib==1.5.0
mlflow==2.22.0
numpy==2.2.5
omegaconf==2.3.0
optuna==4.3.0
pandas==2.2.3
pandera==0.24.0
pyarrow==19.0.1
scikit-learn==1.6.1
